PROD-512 updated tests
