+++ __tests__/actions/answer/answer-question.test.ts
@@ -1,159 +1,260 @@
 import { SaveQuestionRequest, answerQuestion } from "@/app/actions/answer";
 import { getJwtPayload } from "@/app/actions/jwt";
-import { deckSchema } from "@/app/schemas/deck";
 import prisma from "@/app/services/prisma";
-import { QuestionType, Token } from "@prisma/client";
+import { AnswerStatus, QuestionType } from "@prisma/client";
 import { v4 as uuidv4 } from "uuid";
-import { z } from "zod";
 
 jest.mock("@/app/actions/jwt", () => ({
   getJwtPayload: jest.fn(),
 }));
 
-describe("answering already revealed question", () => {
-  const userId = uuidv4();
-  let currentDeckId: number;
-  let currentQuestionId: number;
-  let deckQuestionId: number;
-  let currentQuestionOptions: any;
-  const currentDate = new Date("2024-07-20 16:19:30.717");
+jest.mock("@/app/services/prisma", () => ({
+  questionOption: {
+    findMany: jest.fn(),
+  },
+  questionAnswer: {
+    findMany: jest.fn(),
+    createMany: jest.fn(),
+    deleteMany: jest.fn(),
+  },
+  $transaction: jest.fn(),
+}));
 
-  // Dummy deck data
-  const deckMockData = {
-    deck: `deck ${new Date().getTime()}`,
-    description: "Description",
-    footer: "Footer",
-    tagIds: [],
-    stackId: null,
-    revealToken: "Bonk",
-    activeFromDate: new Date(currentDate.setDate(currentDate.getDate() - 1)),
-    revealAtDate: new Date("2024-07-21 16:19:30.717"),
-    revealTokenAmount: 50,
-    revealAtAnswerCount: null,
-    durationMiliseconds: 10,
-    questions: [],
-    imageUrl: "",
-  } as unknown as z.infer<typeof deckSchema>;
+// Mock the return value for question answers (simulating no previous answers)
+(prisma.questionAnswer.findMany as jest.Mock).mockResolvedValue([]);
 
-  const questionMockData = {
-    question: "What is the capital of France?",
-    type: QuestionType.BinaryQuestion,
-    durationMiliseconds: 60000,
-    revealToken: Token.Bonk,
-    revealAtDate: new Date("2024-07-21 16:19:30.717"),
-    revealTokenAmount: 50,
-  };
+// Mock $transaction to track the operations within it
+const deleteManyMock = jest.fn();
+const createManyMock = jest.fn();
 
-  const questionOptionMockData = [
-    {
-      option: "Paris",
-      isCorrect: true,
-      isLeft: true,
+(prisma.$transaction as jest.Mock).mockImplementation(async (callback) => {
+  await callback({
+    questionAnswer: {
+      deleteMany: deleteManyMock,
+      createMany: createManyMock,
     },
-    {
-      option: "London",
-      isCorrect: false,
-      isLeft: false,
-    },
-  ];
+  });
+});
 
-  beforeAll(async () => {
+describe("answerQuestion", () => {
+  let userId: string;
+  let questionId: number;
+  let questionOptionId: number;
+
+  beforeAll(() => {
+    // Set up mock user ID and question details
+    userId = uuidv4();
+    questionId = 123; // Mock question ID
+    questionOptionId = 4449; // Mock question option ID
+
+    // Mock the return value of getJwtPayload to simulate the user context
     (getJwtPayload as jest.Mock).mockReturnValue({
       sub: userId,
     });
-    const deck = await prisma.deck.create({
-      data: {
-        deck: deckMockData.deck,
-        imageUrl: deckMockData.imageUrl,
-        revealAtDate: deckMockData.revealAtDate,
-        revealAtAnswerCount: deckMockData.revealAtAnswerCount,
-        date: deckMockData.date,
-        activeFromDate: deckMockData.activeFromDate,
-        stackId: deckMockData.stackId,
-        description: deckMockData.description,
-        footer: deckMockData.footer,
-        heading: deckMockData.heading,
-      },
-    });
-    currentDeckId = deck.id;
-
-    const question = await prisma.question.create({
-      data: {
-        question: questionMockData.question,
-        type: questionMockData.type,
-        revealToken: questionMockData.revealToken,
-        revealTokenAmount: questionMockData.revealTokenAmount,
-        revealAtDate: questionMockData.revealAtDate,
-        durationMiliseconds: questionMockData.durationMiliseconds,
-        stackId: null,
-        revealAtAnswerCount: null,
-        imageUrl: "",
-      },
-    });
-
-    currentQuestionId = question.id;
-
-    await prisma.questionOption.createMany({
-      data: questionOptionMockData.map((qo) => ({
-        ...qo,
-        questionId: question.id,
-      })),
-    });
-
-    // Fetch the newly created options
-    const createdOptions = await prisma.questionOption.findMany({
-      where: { questionId: question.id },
-    });
-
-    currentQuestionOptions = createdOptions;
-
-    const deckQuestion = await prisma.deckQuestion.create({
-      data: {
-        deckId: currentDeckId,
-        questionId: currentQuestionId,
-      },
-    });
-
-    deckQuestionId = deckQuestion.id;
   });
 
-  afterAll(async () => {
-    await prisma.deckQuestion.delete({
-      where: {
-        id: deckQuestionId,
+  it("should allow a user to answer a question once", async () => {
+    (prisma.questionOption.findMany as jest.Mock).mockResolvedValue([
+      {
+        id: 4449,
+        option: "Yes",
+        isCorrect: false,
+        calculatedIsCorrect: null,
+        calculatedAveragePercentage: null,
+        calculatedPercentageOfSelectedAnswers: null,
+        isLeft: true,
+        createdAt: new Date("2024-11-20T13:05:20.343Z"),
+        updatedAt: new Date("2024-11-20T13:06:50.463Z"),
+        questionId: questionId,
+        question: {
+          id: questionId,
+          question:
+            "Should we attempt to resurrect extinct species through cloning?",
+          durationMiliseconds: 60000,
+          type: QuestionType.BinaryQuestion,
+          revealToken: "Bonk",
+          revealTokenAmount: 5000,
+          revealAtDate: new Date("2024-11-24T00:00:00.000Z"),
+          revealAtAnswerCount: null,
+          imageUrl: "",
+          createdAt: new Date("2024-11-20T13:05:20.343Z"),
+          updatedAt: new Date("2024-11-20T13:06:50.463Z"),
+          stackId: 36,
+        },
       },
-    });
-    await prisma.deck.delete({
-      where: {
-        id: currentDeckId,
+      {
+        id: 4450,
+        option: "No",
+        isCorrect: false,
+        calculatedIsCorrect: null,
+        calculatedAveragePercentage: null,
+        calculatedPercentageOfSelectedAnswers: null,
+        isLeft: false,
+        createdAt: new Date("2024-11-20T13:05:20.343Z"),
+        updatedAt: new Date("2024-11-20T13:06:50.463Z"),
+        questionId: questionId,
+        question: {
+          id: questionId,
+          question:
+            "Should we attempt to resurrect extinct species through cloning?",
+          durationMiliseconds: 60000,
+          type: QuestionType.BinaryQuestion,
+          revealToken: "Bonk",
+          revealTokenAmount: 5000,
+          revealAtDate: new Date("2024-11-24T00:00:00.000Z"),
+          revealAtAnswerCount: null,
+          imageUrl: "",
+          createdAt: new Date("2024-11-20T13:05:20.343Z"),
+          updatedAt: new Date("2024-11-20T13:06:50.463Z"),
+          stackId: 36,
+        },
       },
-    });
-    await prisma.questionOption.deleteMany({
-      where: {
-        questionId: currentQuestionId,
-      },
-    });
-    await prisma.question.delete({
-      where: {
-        id: currentQuestionId,
-      },
-    });
-  });
+    ]);
 
-  it("should throw an error if the user tries to answer a revealed question", async () => {
+    (prisma.questionAnswer.findMany as jest.Mock).mockResolvedValue([
+      {
+        id: 1200,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+        questionOptionId: 4449,
+        userId,
+        percentage: null,
+        status: AnswerStatus.Viewed,
+        selected: false,
+        timeToAnswer: null,
+      },
+      {
+        id: 1201,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+        questionOptionId: 4450,
+        userId,
+        percentage: null,
+        status: AnswerStatus.Viewed,
+        selected: false,
+        timeToAnswer: null,
+      },
+    ]);
+
     const request: SaveQuestionRequest = {
-      questionId: currentQuestionId,
-      questionOptionId: currentQuestionOptions[0].id,
+      questionId,
+      questionOptionId,
+      percentageGiven: 50,
+      timeToAnswerInMiliseconds: 5000,
+    };
+
+    await answerQuestion(request);
+
+    expect(deleteManyMock).toHaveBeenCalledWith({
+      where: {
+        questionOption: {
+          questionId,
+        },
+        userId,
+      },
+    });
+
+    expect(createManyMock).toHaveBeenCalled();
+  });
+
+  it("should not allow a user to answer the same question twice", async () => {
+    (prisma.questionOption.findMany as jest.Mock).mockResolvedValue([
+      {
+        id: 4449,
+        option: "Yes",
+        isCorrect: false,
+        calculatedIsCorrect: null,
+        calculatedAveragePercentage: null,
+        calculatedPercentageOfSelectedAnswers: null,
+        isLeft: true,
+        createdAt: new Date("2024-11-20T13:05:20.343Z"),
+        updatedAt: new Date("2024-11-20T13:06:50.463Z"),
+        questionId: questionId,
+        question: {
+          id: questionId,
+          question:
+            "Should we attempt to resurrect extinct species through cloning?",
+          durationMiliseconds: 60000,
+          type: QuestionType.BinaryQuestion,
+          revealToken: "Bonk",
+          revealTokenAmount: 5000,
+          revealAtDate: new Date("2024-11-24T00:00:00.000Z"),
+          revealAtAnswerCount: null,
+          imageUrl: "",
+          createdAt: new Date("2024-11-20T13:05:20.343Z"),
+          updatedAt: new Date("2024-11-20T13:06:50.463Z"),
+          stackId: 36,
+        },
+      },
+      {
+        id: 4450,
+        option: "No",
+        isCorrect: false,
+        calculatedIsCorrect: null,
+        calculatedAveragePercentage: null,
+        calculatedPercentageOfSelectedAnswers: null,
+        isLeft: false,
+        createdAt: new Date("2024-11-20T13:05:20.343Z"),
+        updatedAt: new Date("2024-11-20T13:06:50.463Z"),
+        questionId: questionId,
+        question: {
+          id: questionId,
+          question:
+            "Should we attempt to resurrect extinct species through cloning?",
+          durationMiliseconds: 60000,
+          type: QuestionType.BinaryQuestion,
+          revealToken: "Bonk",
+          revealTokenAmount: 5000,
+          revealAtDate: new Date("2024-11-24T00:00:00.000Z"),
+          revealAtAnswerCount: null,
+          imageUrl: "",
+          createdAt: new Date("2024-11-20T13:05:20.343Z"),
+          updatedAt: new Date("2024-11-20T13:06:50.463Z"),
+          stackId: 36,
+        },
+      },
+    ]);
+
+    (prisma.questionAnswer.findMany as jest.Mock).mockResolvedValue([
+      {
+        id: 1202,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+        questionOptionId,
+        userId,
+        percentage: null,
+        status: AnswerStatus.Submitted,
+        selected: true,
+        timeToAnswer: 5000,
+      },
+      {
+        id: 1203,
+        createdAt: new Date(),
+        updatedAt: new Date(),
+        questionOptionId: 4450,
+        userId,
+        percentage: null,
+        status: AnswerStatus.Submitted,
+        selected: false,
+        timeToAnswer: 5000,
+      },
+    ]);
+
+    const request: SaveQuestionRequest = {
+      questionId,
+      questionOptionId,
       percentageGiven: 50,
       timeToAnswerInMiliseconds: 5000,
     };
 
     try {
-      await answerQuestion(request); // Call the method under test
+      await answerQuestion(request);
     } catch (error: any) {
       expect(error.message).toBe(
-        `Question with id: ${request.questionId} does not exist or it is revealed and cannot be answered.`,
+        `User with id: ${userId} has already answered question with id: ${questionId}`,
       );
     }
+
+    expect(prisma.questionAnswer.createMany).not.toHaveBeenCalled();
   });
 });
